pyproject.toml
[tool.poetry]
name = "mht-unarchive"
version = "0.1.0"
description = "Python program to extract from `mht` (Microsoft HTML archive) to html and image files"
authors = ["Ernest Prabhakar <ernest@quiltdata.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.1"

[tool.poetry.group.dev.dependencies]
watchdog = "^2.2.1"
pytest-watch = "^4.2.0"
flake8 = "^6.0.0"

[tool.poetry.scripts]
mht-unarchive = "extract:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
